No issue: Fix dynamic theming to accommodate lazily inflated readerview
