+++ app/src/main/java/org/mozilla/fenix/browser/BrowserFragment.kt
@@ -9,10 +9,7 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.RadioButton
 import androidx.coordinatorlayout.widget.CoordinatorLayout
-import androidx.core.content.ContextCompat
 import androidx.transition.TransitionInflater
 import com.google.android.material.snackbar.Snackbar
 import kotlinx.android.synthetic.main.fragment_browser.*
@@ -100,12 +97,6 @@ class BrowserFragment : BaseBrowserFragment(), UserInteractionHandler {
                 view = view
             )
 
-            if ((activity as HomeActivity).browsingModeManager.mode.isPrivate) {
-                // We need to update styles for private mode programmatically for now:
-                // https://github.com/mozilla-mobile/android-components/issues/3400
-                themeReaderViewControlsForPrivateMode(view.readerViewControlsBar)
-            }
-
             consumeFrom(browserFragmentStore) {
                 browserToolbarView.update(it)
             }
@@ -186,36 +177,6 @@ class BrowserFragment : BaseBrowserFragment(), UserInteractionHandler {
                     gravity = getAppropriateLayoutGravity()
                 )
             nav(R.id.browserFragment, directions)
-        }
-    }
-
-    private fun themeReaderViewControlsForPrivateMode(view: View) = with(view) {
-        listOf(
-            R.id.mozac_feature_readerview_font_size_decrease,
-            R.id.mozac_feature_readerview_font_size_increase
-        ).map {
-            findViewById<Button>(it)
-        }.forEach {
-            it.setTextColor(
-                ContextCompat.getColorStateList(
-                    context,
-                    R.color.readerview_private_button_color
-                )
-            )
-        }
-
-        listOf(
-            R.id.mozac_feature_readerview_font_serif,
-            R.id.mozac_feature_readerview_font_sans_serif
-        ).map {
-            findViewById<RadioButton>(it)
-        }.forEach {
-            it.setTextColor(
-                ContextCompat.getColorStateList(
-                    context,
-                    R.color.readerview_private_radio_color
-                )
-            )
         }
     }
